refactor: retype GetPermissionLevelByTelegramIdAsync to PermissionLevel?
Delete Data/PermissionLevel.cs duplicate; repository now returns the
Core enum instead of int?, giving callers type safety. Data DTOs stay
enum-free (int columns) per the Data project's dependency rule. Widen
UserRecordDto and InviteRecordDto permission fields to int; drop the
HasConversion<int> EF config entries now that the property type is
already int. Update all callers (CommandRouter, WelcomeBypassResolver,
UserRepository, InviteRepository, mappings) and fix test mocks to
return PermissionLevel? instead of int?.

+++ TelegramGroupsAdmin.Telegram/Services/BotCommands/CommandRouter.cs
@@ -186,7 +186,7 @@ public partial class CommandRouter
         {
             _logger.LogDebug("User {TelegramId} is linked to web app user with {PermissionLevel} permissions (global)",
                 telegramId, permissionLevel.Value);
-            return permissionLevel.Value; // 0=Admin, 1=GlobalAdmin, 2=Owner (global)
+            return (int)permissionLevel.Value; // 0=Admin, 1=GlobalAdmin, 2=Owner (global)
         }
 
         // Check Telegram admin permissions (cached, per-chat only)
